nothing of note

+++ src/test.cpp
@@ -15,7 +15,7 @@ int main(void){
     test();
     std::cout << N;
 
-    inf_int<int8_t> a = 10;
+    inf_int<int> a = 10;
 
     std::cout << a << N;
